[CL-3169] Delete resource
